+++ flash/audio/classification/input.py
@@ -17,9 +17,9 @@ from typing import Any, Callable, Dict, List, Optional, Union
 import numpy as np
 import pandas as pd
 
-from flash.core.data.io.classification_input import ClassificationInput, ClassificationState
-from flash.core.data.io.input import DataKeys
-from flash.core.data.utilities.classification import TargetMode
+from flash.core.data.io.classification_input import ClassificationInputMixin, ClassificationState
+from flash.core.data.io.input import DataKeys, Input
+from flash.core.data.utilities.classification import MultiBinaryTargetFormatter
 from flash.core.data.utilities.data_frame import read_csv, resolve_files, resolve_targets
 from flash.core.data.utilities.paths import filter_valid_files, has_file_allowed_extension, make_dataset, PATH_TYPE
 from flash.core.data.utilities.samples import to_samples
@@ -37,7 +37,7 @@ def spectrogram_loader(filepath: str):
     return data
 
 
-class AudioClassificationInput(ClassificationInput):
+class AudioClassificationInput(Input, ClassificationInputMixin):
     @requires("audio")
     def load_sample(self, sample: Dict[str, Any]) -> Dict[str, Any]:
         h, w = sample[DataKeys.INPUT].shape[-2:]  # H x W
@@ -115,7 +115,11 @@ class AudioClassificationDataFrameInput(AudioClassificationFilesInput):
         result = super().load_data(files, targets)
 
         # If we had binary multi-class targets then we also know the labels (column names)
-        if self.training and self.target_mode is TargetMode.MULTI_BINARY and isinstance(target_keys, List):
+        if (
+            self.training
+            and isinstance(self.target_formatter, MultiBinaryTargetFormatter)
+            and isinstance(target_keys, List)
+        ):
             classification_state = self.get_state(ClassificationState)
             self.set_state(ClassificationState(target_keys, classification_state.num_classes))