+++ flash/core/data/io/classification_input.py
@@ -12,17 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from dataclasses import dataclass
-from functools import lru_cache
 from typing import Any, List, Optional, Sequence
 
-from flash.core.data.io.input import Input
-from flash.core.data.properties import ProcessState
-from flash.core.data.utilities.classification import (
-    get_target_details,
-    get_target_formatter,
-    get_target_mode,
-    TargetFormatter,
-)
+from flash.core.data.properties import ProcessState, Properties
+from flash.core.data.utilities.classification import get_target_formatter, TargetFormatter
 
 
 @dataclass(unsafe_hash=True, frozen=True)
@@ -34,9 +27,9 @@ class ClassificationState(ProcessState):
     num_classes: Optional[int] = None
 
 
-class ClassificationInput(Input):
-    """The ``ClassificationInput`` class provides utility methods for handling classification targets.
-    :class:`~flash.core.data.io.input.Input` objects that extend ``ClassificationInput`` should do the following:
+class ClassificationInputMixin(Properties):
+    """The ``ClassificationInputMixin`` class provides utility methods for handling classification targets.
+    :class:`~flash.core.data.io.input.Input` objects that extend ``ClassificationInputMixin`` should do the following:
 
     * In the ``load_data`` method, include a call to ``load_target_metadata``. This will determine the format of the
       targets and store metadata like ``labels`` and ``num_classes``.
@@ -44,27 +37,31 @@ class ClassificationInput(Input):
       tasks.
     """
 
-    @property
-    @lru_cache(maxsize=None)
-    def target_formatter(self) -> TargetFormatter:
-        """Get the :class:`~flash.core.data.utiltiies.classification.TargetFormatter` to use when formatting
-        targets.
-
-        This property uses ``functools.lru_cache`` so that we only instantiate the formatter once.
-        """
-        classification_state = self.get_state(ClassificationState)
-        return get_target_formatter(self.target_mode, classification_state.labels, classification_state.num_classes)
-
-    def load_target_metadata(self, targets: List[Any]) -> None:
+    def load_target_metadata(
+        self, targets: Optional[List[Any]], target_formatter: Optional[TargetFormatter] = None
+    ) -> None:
         """Determine the target format and store the ``labels`` and ``num_classes``.
 
         Args:
             targets: The list of targets.
+            target_formatter: Optionally provide a :class:`~flash.core.data.utilities.classification.TargetFormatter`
+                rather than inferring from the targets.
         """
-        self.target_mode = get_target_mode(targets)
-        self.multi_label = self.target_mode.multi_label
-        if self.get_state(ClassificationState) is None:
-            self.labels, self.num_classes = get_target_details(targets, self.target_mode)
+        if target_formatter is None and targets is not None:
+            classification_state = self.get_state(ClassificationState)
+            if classification_state is not None:
+                labels, num_classes = classification_state.labels, classification_state.num_classes
+            else:
+                labels, num_classes = None, None
+
+            self.target_formatter = get_target_formatter(targets, labels, num_classes)
+        else:
+            self.target_formatter = target_formatter
+
+        if getattr(self, "target_formatter", None) is not None:
+            self.multi_label = self.target_formatter.multi_label
+            self.labels = self.target_formatter.labels
+            self.num_classes = self.target_formatter.num_classes
             self.set_state(ClassificationState(self.labels, self.num_classes))
 
     def format_target(self, target: Any) -> Any:
@@ -76,4 +73,6 @@ class ClassificationInput(Input):
         Returns:
             The formatted target.
         """
-        return self.target_formatter(target)
+        if getattr(self, "target_formatter", None) is not None:
+            return self.target_formatter(target)
+        return target